refactor(frontend): cookie declarations
