+++ apps/frontend/app/routes/_dashboard.tsx
@@ -50,9 +50,9 @@ import { produce } from "immer";
 import { joinURL } from "ufo";
 import { HiddenLocationInput } from "~/components/common";
 import { redirectIfNotAuthenticated } from "~/lib/api.server";
-import { colorSchemeCookie } from "~/lib/cookies.server";
 import { ApplicationKey, getLot } from "~/lib/generals";
 import {
+	colorSchemeCookie,
 	expectedEnvironmentVariables,
 	getCoreDetails,
 	getUserPreferences,
